refactor(common) : constant와 exception 분리 및 관련 class 저장
