+++ src/main/java/lotto/domain/LottoTicket.java
@@ -4,9 +4,9 @@ import camp.nextstep.edu.missionutils.Randoms;
 
 import java.util.List;
 
-import static lotto.common.NumberConstant.LOTTO_LENGTH;
-import static lotto.common.NumberConstant.LOTTO_MIN_VALUE;
-import static lotto.common.NumberConstant.LOTTO_MAX_VALUE;
+import static lotto.common.constant.NumberConstant.LOTTO_LENGTH;
+import static lotto.common.constant.NumberConstant.LOTTO_MIN_VALUE;
+import static lotto.common.constant.NumberConstant.LOTTO_MAX_VALUE;
 
 public class LottoTicket {
